Add selectors
This commit seperates the selector logic from the component.
THis has several effects:
1. You dont have to repeat yourself (DRY) if you need to access the state in another component
2. If you change your store setup, you only have to touch these functions and all your components will still receive the correct data.
3. You can wrap these in memo from memoize-one if you run into perforamnce issues.
