+++ src/components/SearchBox/SearchBox.js
@@ -3,9 +3,10 @@ import { useDispatch, useSelector } from 'react-redux';
 import './SearchBox.scss';
 import CartBox from '../CartBox/CartBox';
 import { setTextFilter } from '../../actions/filters';
+import { getSearchBoxInput } from '../../selectors/filters';
 
 const SearchBox = () => {
-  const inputValue = useSelector((state) => state.filtersReducer.name);
+  const inputValue = useSelector(getSearchBoxInput);
   const dispatch = useDispatch();
 
   const searchPhone = (e) => {
@@ -17,12 +18,12 @@ const SearchBox = () => {
     <div className="search-box">
       <CartBox />
       <h3>Search Phones</h3>
-      <input 
-      className='search-box--input' 
-      type="text" 
-      value={inputValue} 
-      onChange={searchPhone}
-      placeholder='Search...'
+      <input
+        className='search-box--input'
+        type="text"
+        value={inputValue}
+        onChange={searchPhone}
+        placeholder='Search...'
       />
     </div>
   )
