+++ src/selectors/filters.js
@@ -1,0 +1,9 @@
+
+const getAllFilters = state => state.filters;
+
+const getSearchBoxInput = state => state.filters.name;
+
+export {
+    getAllFilters,
+    getSearchBoxInput
+}
