+++ src/store/configureStore.js
@@ -6,7 +6,7 @@ const configureStore = () => {
   const store = createStore(
     combineReducers({
       phonesReducer,
-      filtersReducer
+      filters: filtersReducer
     }),
     window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
   )
